Switch to different authentification approach and clean up README
For unknown reasons, the instance_exec approach no longer works. Rails picks up the before_action, but says the method is undefined. :-/

I don't know if this is Rails or Ruby breaking things, but either I have switched to a simpler model which allows you to define a parent controller.